restructured application hierarchy
